refactor of model_constructor model_constructor is now net_models each model type has its own python file
