+++ src/net_models/__init__.py
@@ -1,0 +1,7 @@
+from . import binary_classifier
+from . import multiclass_classifier
+
+from .evaluate import evaluate
+from .predict import predict
+from .system_setup import system_setup
+from .train import train
