+++ src/visualization/plot_confusion_matrix.py
@@ -1,9 +1,4 @@
-# from matplotlib.pyplot import subplots, show
-from fnmatch import fnmatch
 import matplotlib.pyplot as plt
-from pandas import DataFrame
-from math import ceil
-import numpy as np
 from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay
 
 
@@ -17,12 +12,11 @@ def plot_confusion_matrix(true_labels, predicted_labels, class_names):
         class_names (list): List of class names for the matrix.
     """
 
-    cm = confusion_matrix(
-        true_labels, predicted_labels
-    )  # labels=range(len(class_names)))
+    cm = confusion_matrix(true_labels, predicted_labels)
     disp = ConfusionMatrixDisplay(confusion_matrix=cm, display_labels=class_names)
 
-    plt.figure(figsize=(8, 8))
+    plt.figure(figsize=(14, 14))
     disp.plot(cmap="Blues", values_format="d")
     plt.title("Confusion Matrix")
+    plt.tight_layout()
     plt.show()
